requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.4.0=pypi_0
altair=5.0.1=pypi_0
anyio=3.5.0=py38haa95532_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38h2bbff1b_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
attrs=22.1.0=py38haa95532_0
backcall=0.2.0=pyhd3eb1b0_0
backports-zoneinfo=0.2.1=pypi_0
beautifulsoup4=4.12.2=py38haa95532_0
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.6.2=pypi_0
ca-certificates=2023.05.30=haa95532_0
cachetools=5.3.1=pypi_0
certifi=2023.5.7=pypi_0
cffi=1.15.1=py38h2bbff1b_3
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
colorama=0.4.6=py38haa95532_0
comm=0.1.2=py38haa95532_0
contourpy=1.0.7=pypi_0
cycler=0.11.0=pypi_0
debugpy=1.5.1=py38hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py38haa95532_0
executing=0.8.3=pyhd3eb1b0_0
flatbuffers=23.5.26=pypi_0
fonttools=4.39.4=pypi_0
gast=0.4.0=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
google-auth=2.19.1=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.54.2=pypi_0
h5py=3.8.0=pypi_0
idna=3.4=py38haa95532_0
importlib-metadata=6.0.0=py38haa95532_0
importlib_metadata=6.0.0=hd3eb1b0_0
importlib_resources=5.2.0=pyhd3eb1b0_1
ipykernel=6.19.2=py38hd4e2768_0
ipython=8.12.0=py38haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jax=0.4.12=pypi_0
jedi=0.18.1=py38haa95532_1
jinja2=3.1.2=py38haa95532_0
joblib=1.2.0=pypi_0
jsonschema=4.17.3=py38haa95532_0
jupyter_client=8.1.0=py38haa95532_0
jupyter_core=5.3.0=py38haa95532_0
jupyter_events=0.6.3=py38haa95532_0
jupyter_server=2.5.0=py38haa95532_0
jupyter_server_terminals=0.4.4=py38haa95532_0
jupyterlab_pygments=0.1.2=py_0
keras=2.12.0=pypi_0
kiwisolver=1.4.4=pypi_0
libclang=16.0.0=pypi_0
libffi=3.4.4=hd77b12b_0
libiconv=1.16=h2bbff1b_2
libsodium=1.0.18=h62dcd97_0
libxml2=2.10.3=h0ad7f3c_0
libxslt=1.1.37=h2bbff1b_0
lxml=4.9.2=py38h2bbff1b_0
markdown=3.4.3=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.1=py38h2bbff1b_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=py38haa95532_0
mdurl=0.1.2=pypi_0
mistune=0.8.4=py38he774522_1000
ml-dtypes=0.2.0=pypi_0
nb_conda=2.2.1=py38_1
nb_conda_kernels=2.3.1=py38haa95532_0
nbclassic=0.5.5=py38haa95532_0
nbclient=0.5.13=py38haa95532_0
nbconvert=6.5.4=py38haa95532_0
nbformat=5.7.0=py38haa95532_0
nest-asyncio=1.5.6=py38haa95532_0
notebook=6.5.4=py38haa95532_0
notebook-shim=0.2.2=py38haa95532_0
numpy=1.23.5=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.7.0.72=pypi_0
openssl=1.1.1t=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
packaging=23.0=py38haa95532_0
pandas=2.0.2=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.5.0=pypi_0
pip=23.0.1=py38haa95532_0
pkgutil-resolve-name=1.3.10=py38haa95532_0
platformdirs=2.5.2=py38haa95532_0
prometheus_client=0.14.1=py38haa95532_0
prompt-toolkit=3.0.36=py38haa95532_0
protobuf=4.23.2=pypi_0
psutil=5.9.0=py38h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=12.0.1=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydeck=0.8.1b0=pypi_0
pygments=2.15.1=py38haa95532_1
pympler=1.0.1=pypi_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.18.0=py38h196d8e1_0
python=3.8.16=h6244533_3
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py38haa95532_0
python-json-logger=2.0.7=py38haa95532_0
pytz=2023.3=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pywin32=305=py38h2bbff1b_0
pywinpty=2.0.10=py38h5da7b33_0
pyyaml=6.0=py38h2bbff1b_1
pyzmq=25.0.2=py38hd77b12b_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rfc3339-validator=0.1.4=py38haa95532_0
rfc3986-validator=0.1.1=py38haa95532_0
rich=13.4.2=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=67.8.0=py38haa95532_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sniffio=1.2.0=py38haa95532_1
soupsieve=2.4=py38haa95532_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
streamlit=1.23.1=pypi_0
tenacity=8.2.2=pypi_0
tensorboard=2.12.3=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorflow=2.12.0=pypi_0
tensorflow-estimator=2.12.0=pypi_0
tensorflow-intel=2.12.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.3.0=pypi_0
terminado=0.17.1=py38haa95532_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.2.1=py38haa95532_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
tornado=6.2=py38h2bbff1b_0
tqdm=4.65.0=pypi_0
traitlets=5.7.1=py38haa95532_0
typing-extensions=4.5.0=py38haa95532_0
typing_extensions=4.5.0=py38haa95532_0
tzdata=2023.3=pypi_0
tzlocal=4.3=pypi_0
urllib3=1.26.16=pypi_0
validators=0.20.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=3.0.0=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
websocket-client=0.58.0=py38haa95532_4
werkzeug=2.3.6=pypi_0
wheel=0.38.4=py38haa95532_0
winpty=0.4.3=4
wrapt=1.14.1=pypi_0
yaml=0.2.5=he774522_0
zeromq=4.3.4=hd77b12b_0
zipp=3.11.0=py38haa95532_0
zlib=1.2.13=h8cc25b3_0
